+++ onedrive-backup/Contracts/AddonOptions.cs
@@ -1,4 +1,5 @@
 ﻿using Newtonsoft.Json;
+using System.Text.Json.Serialization;
 using JsonIgnoreAttribute = Newtonsoft.Json.JsonIgnoreAttribute;
 
 namespace hassio_onedrive_backup.Contracts
@@ -32,10 +33,54 @@ namespace hassio_onedrive_backup.Contracts
         [JsonProperty("hass_api_timeout_minutes")]
         public int HassAPITimeoutMinutes { get; set; }
 
+        [JsonProperty("exclude_media_folder")]
+        public bool ExcludeMediaFolder { get; set; }
+
+        [JsonProperty("exclude_ssl_folder")]
+        public bool ExcludeSSLFolder { get; set; }
+
+        [JsonProperty("exclude_share_folder")]
+        public bool ExcludeShareFolder { get; set; }
+
+        [JsonProperty("exclude_local_addons_folder")]
+        public bool ExcludeLocalAddonsFolder { get; set; }
+
         [JsonIgnore]
         public float BackupIntervalHours => BackupIntervalDays * 24;
 
         [JsonIgnore]
         public string BackupNameSafe => string.IsNullOrEmpty(BackupName) ? "hass_backup" : BackupName;
+
+        [JsonIgnore]
+        public bool IsPartialBackup => ExcludeLocalAddonsFolder || ExcludeMediaFolder || ExcludeShareFolder || ExcludeSSLFolder;
+
+        public List<string> IncludedFolderList
+        {
+            get
+            {
+                List<string> folders = new List<string>();
+                if (!ExcludeLocalAddonsFolder)
+                {
+                    folders.Add("addons/local");
+                }
+
+                if (!ExcludeMediaFolder)
+                {
+                    folders.Add("media");
+                }
+
+                if (!ExcludeShareFolder)
+                {
+                    folders.Add("share");
+                }
+
+                if (!ExcludeSSLFolder)
+                {
+                    folders.Add("ssl");
+                }
+
+                return folders;
+            }
+        }
     }
 }