+++ onedrive-backup/Graph/IGraphHelper.cs
@@ -12,8 +12,8 @@ namespace hassio_onedrive_backup.Graph
 
         Task<bool> UploadFileAsync(string filePath, DateTime date, string? destinationFileName = null, Action<int>? progressCallback = null);
 
-        Task<string?> DownloadFileAsync(string fileName, Action<int>? progressCallback);
+        Task<string?> DownloadFileAsync(string fileName, Action<int?>? progressCallback);
 
-        Task<int?> GetFreeSpaceInGB();
+        Task<double?> GetFreeSpaceInGB();
     }
 }